feat: compute task level
